AP-2599: refactor ExhaustedFailureMessage
Move it into a sidekiq namespace to allow multiple messages
Update calls to it
