+++ app/workers/hmrc/base_worker.rb
@@ -7,7 +7,7 @@ module HMRC
     MAX_RETRIES = 10
     sidekiq_options retry: MAX_RETRIES
     sidekiq_retries_exhausted do |msg, _ex|
-      Sentry.capture_message ExhaustedFailureMessage.call(msg)
+      Sentry.capture_message Sidekiq::ExhaustedFailureMessage.call(msg)
     end
 
     def perform(hmrc_response_id)
